worker/requirements.txt
flask_sqlalchemy
psycopg2-binary
shortuuid
pika
pandas
numpy
torch
torch-geometric
scikit-learn
